politisense/requirements.txt
# PolitiSense Agent Dependencies
# Install with: pip install -r requirements.txt

# Vector store
chromadb>=0.4.22

# Embeddings (runs locally, no API key needed)
sentence-transformers>=2.2.2

# Anthropic Claude API
anthropic>=0.20.0

# LangGraph — agent reasoning loop + memory
langgraph>=0.2.0

# Market data — yfinance (Stage 4)
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# Load API keys from .env file
python-dotenv>=1.0.0

# Progress bar
tqdm>=4.66.0

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0
